backend/tests/requirements_test.txt
# Test requirements for MetalQMS
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==24.0.0
coverage==7.4.3
model-bakery==1.17.0
freezegun==1.4.0
responses==0.25.0
hypothesis==6.98.15